ImarisCachedProbabilitiesImg represents probability distribution over channels
It is assumed that pixel values are in the range [0,1].
For each pixel, values across all channels should sum to 1.
Channel 0 represents "background" and is not replicated to Imaris. Instead,
it is reconstructed from the other channels when reading back cells.

If the Imaris image is uint8 or uint16, [0,1] is scaled to the full range
of the respective datatype.